+++ src/main/java/com/shadrin/properties/AccountProperties.java
@@ -3,29 +3,31 @@ package com.shadrin.properties;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 /**
  * Присваивает по дефолту сумму пи создании счета и комиссию за перевод
  */
 @Component
 public class AccountProperties {
 
-    private final double defaultAmount;
+    private final BigDecimal defaultAmount;
 
-    private final double transferCommission;
+    private final BigDecimal transferCommission;
 
     public AccountProperties(
-            @Value("${account.default-amount:1000.0}") double defaultAmount,
-            @Value("${account.transfer-commission:0.015}") double transferCommission) {
+            @Value("${account.default-amount:1000.0}") BigDecimal defaultAmount,
+            @Value("${account.transfer-commission:0.015}") BigDecimal transferCommission) {
         this.defaultAmount = defaultAmount;
         this.transferCommission = transferCommission;
     }
 
 
-    public double getDefaultAmount() {
+    public BigDecimal getDefaultAmount() {
         return defaultAmount;
     }
 
-    public double getTransferCommission() {
+    public BigDecimal getTransferCommission() {
         return transferCommission;
     }
 }
